fix(audit): independent re-audit — sales/ICP runtime fixes, auth/SSE hardening, clippy+CI
Fixes verified bugs across store, worker, crawl, insights, poi, api, and the
web client. Adds regression tests, migration 044, and a Woodpecker CI pipeline.
All gates pass locally: fmt, clippy -D warnings, 4147 tests, release build.

See docs/CHANGELOG.md (B359-B380) for the full list.

+++ crates/worker/src/job_execution/triage.rs
@@ -68,7 +68,7 @@ pub(crate) async fn run_triage_processing(_kind: &JobKind, store: &Arc<PgStore>)
         // Persist scores (composite_score is computed internally by batch_update_scores)
         match queue.batch_update_scores(score_entries).await {
             Ok(count) => {
-                run.succeed(count as u64, &format!("scored {} items", unscored.len()));
+                run.succeed(count, &format!("scored {} items", unscored.len()));
             }
             Err(e) => {
                 run.fail(&format!("failed to persist scores: {e}"));
